+++ frontend/src/container/InfraMonitoringK8sV2/Base/K8sHeader.tsx
@@ -234,7 +234,6 @@ function K8sHeader<TData>({
 							allowClear
 							maxTagCount="responsive"
 							placeholder="Search for attribute"
-							style={{ width: '100%' }}
 							options={groupByOptions}
 							onChange={handleGroupByChange}
 						/>
