+++ frontend/src/container/InfraMonitoringK8sV2/commonUtils.tsx
@@ -4,33 +4,6 @@ import { InframonitoringtypesPodCountsByPhaseDTO } from 'api/generated/services/
 import { StatusCountItem } from './components/GroupedStatusCounts';
 
 /**
- * Formats age in milliseconds to a human-readable string (e.g., "5d 3h", "2h 30m", "45s")
- */
-export function formatAge(ms: number): string {
-	if (ms < 0 || Number.isNaN(ms) || !Number.isFinite(ms)) {
-		return '-';
-	}
-
-	const seconds = Math.floor(ms / 1000);
-	const minutes = Math.floor(seconds / 60);
-	const hours = Math.floor(minutes / 60);
-	const days = Math.floor(hours / 24);
-
-	if (days > 0) {
-		const remainingHours = hours % 24;
-		return remainingHours > 0 ? `${days}d ${remainingHours}h` : `${days}d`;
-	}
-	if (hours > 0) {
-		const remainingMinutes = minutes % 60;
-		return remainingMinutes > 0 ? `${hours}h ${remainingMinutes}m` : `${hours}h`;
-	}
-	if (minutes > 0) {
-		return `${minutes}m`;
-	}
-	return `${seconds}s`;
-}
-
-/**
  * Converts size in bytes to a human-readable string with appropriate units
  */
 export function formatBytes(bytes: number, decimals = 2): string {
